Remove \printlist{organization} from driver
The unified style sheet guidelines never say anything about putting the
organization that hosted a conference in the citation.

Moreover, given that the unified style sheet guidelines want the
conference title to go into the maintitle/booktitle field and given that
the title of a conference is, in many cases, similar, if not identical,
to the name of the organization, it seems unlikely that this field will
ever be needed by this style.